Refactor SDFGrid helpers into focused modules
